feat: use task to charge subscriptions
