+++ packages/server/src/tasks/index.ts
@@ -1,0 +1,5 @@
+import { chargeSubscription } from './charge_subscription';
+
+export const taskFunctions = {
+  charge_subscription: chargeSubscription,
+};
